Moved base log analyzer to new processor class
